refactor!: rename `luarocks` to `luarocks-nvim`
